feat: modified headerModal close
