+++ web/src/components/Header.tsx
@@ -1,14 +1,15 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
-import { useLocation, useNavigate, NavigateOptions } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import styled, { DefaultTheme } from 'styled-components';
 
 import DarkModeIcon from '../assets/darkmode.webp';
 import LightModeIcon from '../assets/lightmode.webp';
 import { Icons } from '../assets/svg';
-import { setLogin, setSubPageOpen } from '../reducers/header';
+import { setSubPageOpen } from '../reducers/header';
 import { RootState } from '../reducers/index';
+import HeaderModal from './Modal/HeaderModal';
 
 const HeaderContainer = styled.header<{ isTransitionEnabled: boolean }>`
   position: sticky;
@@ -176,42 +177,6 @@ const LoginInformation = styled.div`
   }
 `;
 
-const SubPageContainer = styled.div`
-  display: flex;
-  transition: opacity 0.4s ease-in-out;
-  position: absolute;
-  top: 100%;
-  right: 1%;
-  flex-direction: column;
-  background-color: ${(props) => props.theme.colors.header};
-  z-index: 10;
-  padding: 10px;
-  box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.2);
-  height: auto;
-  overflow-y: auto;
-  border: ${(props) => props.theme.borderResultList};
-  border-radius: ${(props) => props.theme.borderRadius};
-  & > * {
-    margin-bottom: 10px;
-  }
-`;
-
-const SubPageButton = styled.button`
-  margin-top: 1vh;
-  border: none;
-  padding: 1.5vh;
-  border-radius: ${(props) => props.theme.borderRadius};
-  font-size: 2vh;
-  font-weight: 700;
-  color: ${(props) => props.theme.colors.text};
-  background-color: ${(props) => props.theme.colors.button};
-  cursor: pointer;
-
-  &:hover {
-    background-color: ${(props) => props.theme.colors.btnhover};
-  }
-`;
-
 const SaveUserInformationButton = styled.div`
   margin: 1vw;
   display: flex;
@@ -243,33 +208,9 @@ export default function Header({ theme, toggleTheme }: HeaderProps) {
   const isSubPageOpen = useSelector((state: RootState) => state.header.isSubPageOpen);
   const dispatch = useDispatch();
 
-  useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      if (isSubPageOpen === true) {
-        event.preventDefault(); // 경고 메시지 표시
-        dispatch(setSubPageOpen(false));
-      }
-    };
-    window.addEventListener('beforeunload', handleBeforeUnload);
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-    };
-  }, []);
-
   const handleClick = () => {
     setIsTransitionEnabled(true);
     toggleTheme();
-  };
-
-  const logoutClick = () => {
-    dispatch(setLogin(!isLogin));
-    dispatch(setSubPageOpen(!isSubPageOpen));
-    navigate('../../../');
-  };
-
-  const navigateMypage = () => {
-    navigate('../../../mypage');
-    dispatch(setSubPageOpen(!isSubPageOpen));
   };
 
   return (
@@ -321,16 +262,7 @@ export default function Header({ theme, toggleTheme }: HeaderProps) {
         </CheckBoxContainer>
       </ButtonContainer>
 
-      {isSubPageOpen && (
-        <SubPageContainer theme={theme}>
-          <SubPageButton onClick={navigateMypage} theme={theme}>
-            마이페이지
-          </SubPageButton>
-          <SubPageButton onClick={logoutClick} theme={theme}>
-            로그아웃
-          </SubPageButton>
-        </SubPageContainer>
-      )}
+      {isSubPageOpen && <HeaderModal theme={theme} />}
     </HeaderContainer>
   );
 }
