+++ web/src/hooks/history.ts
@@ -1,0 +1,3 @@
+import { createBrowserHistory } from 'history';
+
+export const history = createBrowserHistory();
