+++ web/src/hooks/useOnClickOutside.ts
@@ -13,9 +13,9 @@ export const useOnClickOutside = ({ ref, handler }: Props) => {
       }
       handler();
     };
-    document.addEventListener('mousedown', listener);
+    window.addEventListener('mousedown', listener);
     return () => {
-      document.removeEventListener('mousedown', listener);
+      window.removeEventListener('mousedown', listener);
     };
   }, [ref, handler]);
 };
